DAT-797-refactor connect module
