--- app/src/app/connect/connect.component.scss
@@ -1,10 +1,0 @@
-#connection-header {
-  margin-top: 2em;
-  margin-bottom: 2em;
-
-  #connection-logo {
-    margin-top: 2em;
-    margin-bottom: 2em;
-  }
-
-}
