+++ app/src/app/modules/connect/connect.component.scss
@@ -1,0 +1,9 @@
+#connection-header {
+    margin-top: 2em;
+    margin-bottom: 2em;
+
+    #connection-logo {
+        margin-top: 2em;
+        margin-bottom: 2em;
+    }
+}
